+++ src/main/java/duke/command/HelpCommand.java
@@ -1,7 +1,11 @@
-package duke;
+package duke.command;
+
+import duke.task.TaskList;
+import duke.ui.Storage;
+import duke.ui.Ui;
 
 /**
- * Represents a command to exit the program.
+ * Represents a command to display the in-app help tool.
  *
  * @author Sherman Ng Wei Sheng
  */
@@ -9,7 +13,7 @@ public class HelpCommand extends Command {
     private final boolean isExit;
 
     /**
-     * Constructor for the exit command.
+     * Constructor for the help command.
      */
     public HelpCommand() {
         this.isExit = false;
@@ -26,16 +30,15 @@ public class HelpCommand extends Command {
     }
 
     /**
-     * Executes the command to exit the program.
+     * Executes the command to return the content of the help page.
      *
      * @param list TaskList before execution of the command.
      * @param ui Ui object to log the execution of the command.
      * @param storage Storage object that references the path to store the updated list of tasks.
      * @return The string to be printed.
-     * @throws StorageSavingException If exception encountered when storing the list.
      */
     @Override
-    public String execute(TaskList list, Ui ui, Storage storage) throws StorageLoadingException {
+    public String execute(TaskList list, Ui ui, Storage storage) {
         String data = HelpProvider.getContent();
         return ui.printAndReturnMessage(data);
     }
