+++ src/main/java/duke/command/PrintCommand.java
@@ -1,4 +1,8 @@
-package duke;
+package duke.command;
+
+import duke.task.TaskList;
+import duke.ui.Storage;
+import duke.ui.Ui;
 
 /**
  * Represents the command to print the list.
@@ -16,7 +20,7 @@ public class PrintCommand extends Command {
     }
 
     /**
-     * Returns true if the command is a programme terminating command.
+     * Returns true if the command is a program terminating command.
      *
      * @return True if it is a terminating command and false otherwise.
      */
