+++ src/test/java/duke/task/ToDoTest.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
